nocloud-old-code/requirements.txt
alembic
Click
configparser
dominate
Flask
Flask-Bootstrap
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-Testing
Flask-WTF
itsdangerous
Jinja2
libvirt-python
Mako
MarkupSafe
pylint-flask
PyMySQL
python-dateutil
python-editor
six
SQLAlchemy
visitor
Werkzeug
WTForms
pylxd
paramiko
pyyaml
passlib
